+++ R/util.R
@@ -16,6 +16,15 @@ is_osx <- function() {
   if (is.null(x)) y else x
 }
 
+# a la shiny:::is_available
+is_available <- function(package, version = NULL) {
+  installed <- nzchar(system.file(package = package))
+  if (is.null(version)) {
+    return(installed)
+  }
+  installed && isTRUE(utils::packageVersion(package) >= version)
+}
+
 # determine the output file for a pandoc conversion
 pandoc_output_file <- function(input, pandoc_options) {
   to <- strsplit(pandoc_options$to, "[+-]")[[1]][[1]]
@@ -105,7 +114,9 @@ one_string <- function(x) paste(x, collapse = '\n')
 
 # in a future version of yaml, it will disable the evaluation of !expr but we
 # still need it (https://github.com/rstudio/rmarkdown/issues/1387)
-yaml_load <- function(...) yaml::yaml.load(..., eval.expr = TRUE)
+yaml_load <- function(...) {
+  yaml::yaml.load(..., eval.expr = TRUE)
+}
 
 yaml_load_file <- function(input, ...) yaml_load(read_utf8(input), ...)